+++ lib/xapi-stdext-std/listext_test.ml
@@ -17,8 +17,8 @@ let test_list tested_f (name, case, expected) =
   let check () = Alcotest.(check @@ list int) name expected (tested_f case) in
   (name, `Quick, check)
 
-let test_option tested_f (name, case, expected) =
-  let check () = Alcotest.(check @@ option int) name expected (tested_f case) in
+let test_option typ tested_f (name, case, expected) =
+  let check () = Alcotest.(check @@ option typ) name expected (tested_f case) in
   (name, `Quick, check)
 
 let test_chopped_list tested_f (name, case, expected) =
@@ -180,6 +180,61 @@ let test_sub =
   let tests = List.map test specs in
   ("sub", tests)
 
+let test_find_minimum (name, pp, typ, specs) =
+  let test ((cmp, cmp_name), input, expected) =
+    let name = Printf.sprintf "%s of [%s]" cmp_name (pp input) in
+    test_option typ (Listext.find_minimum cmp) (name, input, expected)
+  in
+  let tests = List.map test specs in
+  (Printf.sprintf "find_minimum (%s)" name, tests)
+
+let test_find_minimum_int =
+  let ascending = (Int.compare, "ascending") in
+  let descending = ((fun a b -> Int.compare b a), "descending") in
+  let specs_int =
+    ( "int"
+    , (fun a -> String.concat "; " (List.map string_of_int a))
+    , Alcotest.int
+    , [
+        (ascending, [], None)
+      ; (ascending, [1; 2; 3; 4; 5], Some 1)
+      ; (ascending, [2; 3; 1; 5; 4], Some 1)
+      ; (descending, [], None)
+      ; (descending, [1; 2; 3; 4; 5], Some 5)
+      ; (descending, [2; 3; 1; 5; 4], Some 5)
+      ]
+    )
+  in
+  test_find_minimum specs_int
+
+let test_find_minimum_tuple =
+  let ascending = ((fun (a, _) (b, _) -> Int.compare a b), "ascending") in
+  let descending = ((fun (a, _) (b, _) -> Int.compare b a), "descending") in
+  let specs_tuple =
+    ( "tuple"
+    , (fun a ->
+        String.concat "; "
+          (List.map (fun (a, b) -> "(" ^ string_of_int a ^ ", " ^ b ^ ")") a)
+      )
+    , Alcotest.(pair int string)
+    , [
+        (ascending, [(1, "fst"); (1, "snd")], Some (1, "fst"))
+      ; (descending, [(1, "fst"); (1, "snd")], Some (1, "fst"))
+      ; (ascending, [(1, "fst"); (1, "snd"); (2, "nil")], Some (1, "fst"))
+      ; (descending, [(1, "nil"); (2, "fst"); (2, "snd")], Some (2, "fst"))
+      ]
+    )
+  in
+  test_find_minimum specs_tuple
+
 let () =
   Alcotest.run "Listext"
-    [test_iteri_right; test_take; test_drop; test_chop; test_sub]
+    [
+      test_iteri_right
+    ; test_take
+    ; test_drop
+    ; test_chop
+    ; test_sub
+    ; test_find_minimum_int
+    ; test_find_minimum_tuple
+    ]